code-samples/Python/semantic-kernel-planning-agent/requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
semantic-kernel>=0.9.0
typing-extensions>=4.5.0